requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.7.4

sentence-transformers==2.7.0
chromadb==0.5.0
hnswlib==0.8.0
onnxruntime==1.19.2

pdfplumber==0.11.4
reportlab==4.0.8

openai==1.59.5
